move ClonedInvoiceLinesInfo in InvoiceDetailCloneMapper
